# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vibe Coder Marketplace API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GigType = (typeof GigType)[keyof typeof GigType];

export const GigType = {
  task: "task",
  hourly: "hourly",
  build: "build",
} as const;

export type GigStatus = (typeof GigStatus)[keyof typeof GigStatus];

export const GigStatus = {
  open: "open",
  closed: "closed",
  in_progress: "in_progress",
} as const;

export interface Gig {
  id: number;
  title: string;
  description: string;
  type: GigType;
  status: GigStatus;
  tags: string[];
  requirements?: string | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  hourlyRate?: number | null;
  recordingPath?: string | null;
  publicSlug: string;
  replyCount: number;
  createdAt: string;
  updatedAt: string;
}

export type CreateGigBodyType =
  (typeof CreateGigBodyType)[keyof typeof CreateGigBodyType];

export const CreateGigBodyType = {
  task: "task",
  hourly: "hourly",
  build: "build",
} as const;

export interface CreateGigBody {
  title: string;
  description: string;
  type: CreateGigBodyType;
  tags: string[];
  requirements?: string | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  hourlyRate?: number | null;
  recordingPath?: string | null;
}

export type UpdateGigBodyType =
  (typeof UpdateGigBodyType)[keyof typeof UpdateGigBodyType];

export const UpdateGigBodyType = {
  task: "task",
  hourly: "hourly",
  build: "build",
} as const;

export type UpdateGigBodyStatus =
  (typeof UpdateGigBodyStatus)[keyof typeof UpdateGigBodyStatus];

export const UpdateGigBodyStatus = {
  open: "open",
  closed: "closed",
  in_progress: "in_progress",
} as const;

export interface UpdateGigBody {
  title?: string;
  description?: string;
  type?: UpdateGigBodyType;
  status?: UpdateGigBodyStatus;
  tags?: string[];
  requirements?: string | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  hourlyRate?: number | null;
  recordingPath?: string | null;
}

export interface GigReply {
  id: number;
  gigId: number;
  senderName: string;
  message?: string | null;
  voiceNotePath?: string | null;
  createdAt: string;
}

export interface CreateGigReplyBody {
  senderName: string;
  message?: string | null;
  voiceNotePath?: string | null;
}

export interface Freelancer {
  id: number;
  name: string;
  bio?: string | null;
  avatarPath?: string | null;
  skills: string[];
  tags: string[];
  hourlyRate?: number | null;
  portfolioLinks: string[];
  contactInfo?: string | null;
  notes?: string | null;
  tools: string[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateFreelancerBody {
  name: string;
  bio?: string | null;
  avatarPath?: string | null;
  skills: string[];
  tags: string[];
  hourlyRate?: number | null;
  portfolioLinks: string[];
  contactInfo?: string | null;
  notes?: string | null;
  tools: string[];
}

export interface UpdateFreelancerBody {
  name?: string;
  bio?: string | null;
  avatarPath?: string | null;
  skills?: string[];
  tags?: string[];
  hourlyRate?: number | null;
  portfolioLinks?: string[];
  contactInfo?: string | null;
  notes?: string | null;
  tools?: string[];
}

export type AvailabilitySlotWorkType =
  (typeof AvailabilitySlotWorkType)[keyof typeof AvailabilitySlotWorkType];

export const AvailabilitySlotWorkType = {
  troubleshooting: "troubleshooting",
  coworking: "coworking",
  build: "build",
  any: "any",
} as const;

export interface AvailabilitySlot {
  id: number;
  freelancerId: number;
  freelancerName: string;
  date: string;
  startTime: string;
  endTime: string;
  durationHours: number;
  hourlyRate?: number | null;
  workType: AvailabilitySlotWorkType;
  notes?: string | null;
  isBooked: boolean;
  bookedByNote?: string | null;
  createdAt: string;
}

export type CreateAvailabilitySlotBodyWorkType =
  (typeof CreateAvailabilitySlotBodyWorkType)[keyof typeof CreateAvailabilitySlotBodyWorkType];

export const CreateAvailabilitySlotBodyWorkType = {
  troubleshooting: "troubleshooting",
  coworking: "coworking",
  build: "build",
  any: "any",
} as const;

export interface CreateAvailabilitySlotBody {
  freelancerId: number;
  date: string;
  startTime: string;
  endTime: string;
  durationHours: number;
  hourlyRate?: number | null;
  workType: CreateAvailabilitySlotBodyWorkType;
  notes?: string | null;
}

export interface UpdateAvailabilitySlotBody {
  isBooked?: boolean;
  bookedByNote?: string | null;
  hourlyRate?: number | null;
  notes?: string | null;
}

export interface ShowcaseProject {
  id: number;
  name: string;
  description: string;
  url?: string | null;
  screenshotPath?: string | null;
  tags: string[];
  tools: string[];
  builderName: string;
  upvotes: number;
  createdAt: string;
  updatedAt: string;
}

export interface CreateShowcaseProjectBody {
  name: string;
  description: string;
  url?: string | null;
  screenshotPath?: string | null;
  tags: string[];
  tools: string[];
  builderName: string;
}

export interface UpdateShowcaseProjectBody {
  name?: string;
  description?: string;
  url?: string | null;
  screenshotPath?: string | null;
  tags?: string[];
  tools?: string[];
  builderName?: string;
}

export type TagCategory = (typeof TagCategory)[keyof typeof TagCategory];

export const TagCategory = {
  tool: "tool",
  skill: "skill",
  gig_type: "gig_type",
  general: "general",
} as const;

export interface Tag {
  id: number;
  name: string;
  category: TagCategory;
  createdAt: string;
}

export type CreateTagBodyCategory =
  (typeof CreateTagBodyCategory)[keyof typeof CreateTagBodyCategory];

export const CreateTagBodyCategory = {
  tool: "tool",
  skill: "skill",
  gig_type: "gig_type",
  general: "general",
} as const;

export interface CreateTagBody {
  name: string;
  category: CreateTagBodyCategory;
}

export type DashboardStatsGigsByType = {
  task: number;
  hourly: number;
  build: number;
};

export interface DashboardStats {
  totalGigs: number;
  openGigs: number;
  totalFreelancers: number;
  openSlots: number;
  totalReplies: number;
  totalShowcaseProjects: number;
  gigsByType: DashboardStatsGigsByType;
}

export type ActivityItemType =
  (typeof ActivityItemType)[keyof typeof ActivityItemType];

export const ActivityItemType = {
  gig_created: "gig_created",
  reply_received: "reply_received",
  freelancer_added: "freelancer_added",
  slot_posted: "slot_posted",
  showcase_submitted: "showcase_submitted",
} as const;

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  createdAt: string;
}

export interface RequestUploadUrlBody {
  name: string;
  size: number;
  contentType: string;
}

export interface RequestUploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ListGigsParams = {
  type?: ListGigsType;
  tag?: string;
  status?: ListGigsStatus;
};

export type ListGigsType = (typeof ListGigsType)[keyof typeof ListGigsType];

export const ListGigsType = {
  task: "task",
  hourly: "hourly",
  build: "build",
} as const;

export type ListGigsStatus =
  (typeof ListGigsStatus)[keyof typeof ListGigsStatus];

export const ListGigsStatus = {
  open: "open",
  closed: "closed",
  in_progress: "in_progress",
} as const;

export type ListFreelancersParams = {
  tag?: string;
  search?: string;
};

export type ListAvailabilitySlotsParams = {
  freelancerId?: number;
  from?: string;
  to?: string;
  booked?: boolean;
};

export type ListShowcaseProjectsParams = {
  tag?: string;
};