# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vibe Coder Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /freelancers:
    get:
      operationId: listFreelancers
      tags: [freelancers]
      summary: List all freelancers
      parameters:
        - name: tag
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of freelancers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Freelancer"
    post:
      operationId: createFreelancer
      tags: [freelancers]
      summary: Add a freelancer to the CRM
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFreelancerBody"
      responses:
        "201":
          description: Created freelancer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
  /freelancers/{id}:
    get:
      operationId: getFreelancer
      tags: [freelancers]
      summary: Get a freelancer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Freelancer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
        "404":
          description: Not found
    put:
      operationId: updateFreelancer
      tags: [freelancers]
      summary: Update a freelancer profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFreelancerBody"
      responses:
        "200":
          description: Updated freelancer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Freelancer"
    delete:
      operationId: deleteFreelancer
      tags: [freelancers]
      summary: Remove a freelancer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── AVAILABILITY ─────────────────────────────────────────────────────────────
  /availability:
    get:
      operationId: listAvailabilitySlots
      tags: [availability]
      summary: List availability slots
      parameters:
        - name: freelancerId
          in: query
          schema:
            type: integer
        - name: from
          in: query
          schema:
            type: string
            format: date
        - name: to
          in: query
          schema:
            type: string
            format: date
        - name: booked
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of availability slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AvailabilitySlot"
    post:
      operationId: createAvailabilitySlot
      tags: [availability]
      summary: Post an availability slot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAvailabilitySlotBody"
      responses:
        "201":
          description: Created slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
  /availability/{id}:
    put:
      operationId: updateAvailabilitySlot
      tags: [availability]
      summary: Update an availability slot (e.g. mark as booked)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAvailabilitySlotBody"
      responses:
        "200":
          description: Updated slot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilitySlot"
    delete:
      operationId: deleteAvailabilitySlot
      tags: [availability]
      summary: Delete an availability slot
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── SHOWCASE ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Freelancer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarPath:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          nullable: true
        portfolioLinks:
          type: array
          items:
            type: string
        contactInfo:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tools:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - skills
        - tags
        - portfolioLinks
        - tools
        - createdAt
        - updatedAt
    CreateFreelancerBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarPath:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          nullable: true
        portfolioLinks:
          type: array
          items:
            type: string
        contactInfo:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tools:
          type: array
          items:
            type: string
      required:
        - name
        - skills
        - tags
        - portfolioLinks
        - tools
    UpdateFreelancerBody:
      type: object
      properties:
        name:
          type: string
        bio:
          type: string
          nullable: true
        avatarPath:
          type: string
          nullable: true
        skills:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: string
        hourlyRate:
          type: number
          nullable: true
        portfolioLinks:
          type: array
          items:
            type: string
        contactInfo:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        tools:
          type: array
          items:
            type: string

    # ─── AVAILABILITY ─────────────────────────────────────────────────────────
    AvailabilitySlot:
      type: object
      properties:
        id:
          type: integer
        freelancerId:
          type: integer
        freelancerName:
          type: string
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        durationHours:
          type: number
        hourlyRate:
          type: number
          nullable: true
        workType:
          type: string
          enum: [troubleshooting, coworking, build, any]
        notes:
          type: string
          nullable: true
        isBooked:
          type: boolean
        bookedByNote:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - freelancerId
        - freelancerName
        - date
        - startTime
        - endTime
        - durationHours
        - workType
        - isBooked
        - createdAt
    CreateAvailabilitySlotBody:
      type: object
      properties:
        freelancerId:
          type: integer
        date:
          type: string
          format: date
        startTime:
          type: string
        endTime:
          type: string
        durationHours:
          type: number
        hourlyRate:
          type: number
          nullable: true
        workType:
          type: string
          enum: [troubleshooting, coworking, build, any]
        notes:
          type: string
          nullable: true
      required:
        - freelancerId
        - date
        - startTime
        - endTime
        - durationHours
        - workType
    UpdateAvailabilitySlotBody:
      type: object
      properties:
        isBooked:
          type: boolean
        bookedByNote:
          type: string
          nullable: true
        hourlyRate:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true

    # ─── SHOWCASE ─────────────────────────────────────────────────────────────
...